API commit 11/6
I was able to successfully create and test the API connection for both
my resources (user and animals) I can now successfully create, read,
update and delete both via the CLI.

To do this, I had to work on my ID and ID connections including the
user/creator, owner and the token - that appeared to be the source of
the bulk of my issues.  I troubleshooted with a few peers to confirm.
Additionally, I adjusted my Github connection to the repo (one for the
API, one for the client side) and I connected the server.js router to my
Animal schema which had given me a few issues.  As a next step, I need
to deploy on Heroku and confirm success there before I can begin working
on my frontend.