Final Push - API Side
Attached is my final push pre-presentation for the client side of my
second project.  The goal was to create an environment that was simple
in UI but useful in a real world scenario.  I went with a central-
oriented theme so that the users eyes are not lost when trying to
navigate.  The first thing I took care of setting up was the API side,
then I moved over towards the client side.

The final but I encountered that I am still working on resolving is
completing a DELETE request with my allocated resource.  At the moment,
it generates an Error 422 in the browser while testing.  I am still
researching into why this is happening.  The CSS is still not where I
want it to be, but that's also considered a work in progress.