Final Update for Submittal
In this verison, I updated my README file to contain the link to the
alternate repo, the deployed API and the deployed client-side site.  I
also momentarily updated my animal.js file in an attempt to solve my
Update via deployed issue, but reverted it back to what it was
originally when I confirmed the issue was in the HTML and not the
API-side.